+++ src/components/AuthView/styles.jsx
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const CredentialsBox = styled.div`
+  width: 480px;
+
+  & > p:first-of-type {
+    margin-bottom: 0.5rem;
+  }
+`;
